BlueBerry [Update 2.1]/requirements.txt
discord.py==1.5.1
db-sqlite3==0.0.1
nekos.py==1.0.3
asyncio==3.4.3
db==0.1.1
Colorama
websockets
startup
aiohttp
chardet
psutil